+++ catapult/test/test_util.py
@@ -42,6 +42,9 @@ class TestUtil(catapult.test.TestCase):
     def test_get_desktop_environment(self):
         catapult.util.get_desktop_environment()
 
+    def test_get_monitor(self):
+        assert catapult.util.get_monitor()
+
     def test_get_screen_size(self):
         width, height = catapult.util.get_screen_size()
         assert width and height